Wire polynomials can be evaluated at X.w
For custom gates (like MiMC), we need to access the wires that
feed into the *following* gate in a circuit. This improves the
efficiency of sequential operations. For example, the MiMC gate
evaluates one round of the MiMC hash function. Here, the
output wire `w_o` is used as an *input* wire, that defines the
hash round input. The output wire for the *next* gate is used
as the output wire. This allows us to chain repeated MiMC hash
gates together and allows our MiMC hash round gate to operate
on 4 wire values instead of 3.

Prover/Verifier widget classes now encapsulate
a 'WidgetVersionControl' class, which  can be
queried by the prover/verifier to determine whether the proof
in question requires w_l(X.w), w_r(X.w) or w_o(X.w) to be included
in the proof. Doing so will increase the proof size by one field
element each, but has a marginal effect on prover/verifier efficiency