Added output wire selector into bool widget
For sequential operations, we can create
fan-in-3 gates by using the following gate's
output wire as an input into the current gate.

This creates situations where 3 wires in a single
gate need to be constraint to be boolean.

The new selector in the bool widget enables this. It
also dramatically simplifies circuit optimization, as
all wire types are now symmetric with respect to booleanity
constraints.