Added bitarray class into stdlib
`bitarray` is designed to be a statically-sized container that holds bits.
Functions as a buffer, that data can be moved into/out of, primarily for
the purposes of evaluating hashes.

Added extremely hacky method into sha256 that accepts a bitarray and
returns a bitarray. Is in a very unfinished state.