added left and right shifts to turbo uint
evaluating logical shifts on accumulating quaternary sums poses some
tricky problems.

to ensure we can perform shifts and rotations in a single gate, the
turbo arithmetic widget has been extended, such that if
q_arith(g) = 2, we assume the 3rd and 4th wires are
adjacent quaternary accumulating sums, and extract the high bit.

n.b. this constraint does not validate that the two witnesses are in
fact quaternary accumulating sums, we assume they have been
copied from the outputs of a range constraint